@@ -2,7 +2,7 @@
   "name": "@openzim/libzim",
   "main": "dist/index.js",
   "types": "dist/index.d.js",
-  "version": "3.5.1-dev0",
+  "version": "4.0.0-dev0",
   "description": "Libzim bindings for NodeJS",
   "type": "module",
   "scripts": {
@@ -14,8 +14,8 @@
     "build": "node-gyp rebuild -v && npm run bundle",
     "download": "node ./download-libzim.js",
     "bundle": "node ./bundle-libzim.js",
-    "test": "jest --testPathIgnorePatterns=test/dist.test.ts",
-    "test-mem-leak": "node -r ts-node/register test/makeLargeZim.ts",
+    "test": "jest --testPathIgnorePatterns=test/dist.test.ts --testPathIgnorePatterns=dist/",
+    "test-mem-leak": "npx tsx test/makeLargeZim.ts",
     "test:dist": "npm run build && jest",
     "lint": "npx eslint .",
     "lint:fix": "npx eslint . --fix"
@@ -44,31 +44,31 @@
   "homepage": "https://github.com/openzim/node-libzim#readme",
   "gypfile": true,
   "dependencies": {
-    "@types/bindings": "^1.5.1",
-    "@types/jest": "^29.5.12",
-    "@types/node": "^18.0.6",
-    "axios": "^1.6.0",
+    "@types/bindings": "^1.5.5",
+    "@types/jest": "^30.0.0",
+    "@types/node": "^24.9.2",
+    "axios": "^1.13.1",
     "bindings": "^1.5.0",
-    "dotenv": "^16.0.1",
+    "dotenv": "^17.2.3",
     "exec-then": "^1.3.1",
     "mkdirp": "^3.0.1",
-    "node-addon-api": "^8.0.0",
-    "node-gyp": "^10.1.0",
+    "node-addon-api": "^8.5.0",
+    "node-gyp": "^11.5.0",
     "tqdm": "^2.0.3",
-    "ts-node": "^10.9.1",
-    "tsconfig-paths": "^4.0.0"
+    "ts-node": "^10.9.2",
+    "tsconfig-paths": "^4.2.0"
   },
   "devDependencies": {
-    "@faker-js/faker": "^8.4.1",
-    "@typescript-eslint/eslint-plugin": "^5.59.2",
-    "@typescript-eslint/parser": "^5.59.2",
+    "@faker-js/faker": "^10.1.0",
+    "@typescript-eslint/eslint-plugin": "^5.62.0",
+    "@typescript-eslint/parser": "^5.62.0",
     "eslint": "^8.57.0",
-    "eslint-config-prettier": "^8.10.0",
-    "eslint-plugin-prettier": "^5.1.3",
-    "nyc": "^17.0.0",
-    "prettier": "^3.3.2",
-    "ts-jest": "^29.1.5",
-    "typescript": "^5.1.6"
+    "eslint-config-prettier": "^8.10.2",
+    "eslint-plugin-prettier": "^5.5.4",
+    "nyc": "^17.1.0",
+    "prettier": "^3.6.2",
+    "ts-jest": "^29.4.5",
+    "typescript": "^5.9.3"
   },
   "jest": {
     "preset": "ts-jest/presets/js-with-ts"
